Change: make build more flexible
In order to fork a scanner-lab repository and use custom images it is
possible to override the ghcr.io address by setting
`GITHUB_REPOSITORY_OWNER` environment variable and to use a different
registry altogether set the `SL_C_REGISTRY` environment variable.

To setup the workflow you need to set the `GREENBONE_BOT` secret in your
repository to your username and create a token and store it as
`GREENBONE_BOT_TOKEN`. The container.yml definition will then pick up
the secrets so that the images are stored as a repository package.